[DPEDE-2326] Rename task
