+++ tasks/update-dist-files.js
